Moved SubjectCache to another package
Part of #588
